Move QgsFeatureFilter and QgsFeatureFilterProviderGroup from server into core
