+++ src/core/qgsfeaturefilter.cpp
