+++ src/core/qgsfeaturefilterprovidergroup.cpp
