+++ src/core/qgsfeaturefilterprovidergroup.h
@@ -1,6 +1,6 @@
 /***************************************************************************
-                              qgsfeaturefilter.h
-                              ------------------
+                       qgsfeaturefilterprovidergroup.h
+                       -------------------------------
   begin                : 26-10-2017
   copyright            : (C) 2017 by Patrick Valsecchi
   email                : patrick dot valsecchi at camptocamp dot com
@@ -15,29 +15,27 @@
  *                                                                         *
  ***************************************************************************/
 
-#ifndef QGSFEATUREFILTER_H
-#define QGSFEATUREFILTER_H
+#ifndef QGSFEATUREFILTERPROVIDERGROUP_H
+#define QGSFEATUREFILTERPROVIDERGROUP_H
 
 #include "qgsfeaturefilterprovider.h"
-#include "qgis_server.h"
+#include "qgis_core.h"
 
-#include <QMap>
-
-class QgsExpression;
+#include <QList>
 
 /**
- * \ingroup server
- * \class QgsFeatureFilter
- * \brief A feature filter provider allowing to set filter expressions on a per-layer basis.
+ * \ingroup core
+ * \class QgsFeatureFilterProviderGroup
+ * \brief A filter filter provider grouping several filter providers.
  */
-class SERVER_EXPORT QgsFeatureFilter : public QgsFeatureFilterProvider
+class CORE_EXPORT QgsFeatureFilterProviderGroup : public QgsFeatureFilterProvider
 {
   public:
     //! Constructor
-    QgsFeatureFilter() = default;
+    QgsFeatureFilterProviderGroup() = default;
 
     /**
-     * Filter the features of the layer
+     * Filter the features of the layer.
      * \param layer the layer to control
      * \param filterFeatures the request to fill
      */
@@ -52,14 +50,15 @@ class SERVER_EXPORT QgsFeatureFilter : public QgsFeatureFilterProvider
     QgsFeatureFilterProvider *clone() const override SIP_FACTORY;
 
     /**
-     * Set a filter for the given layer.
-     * \param layer the layer to filter
-     * \param expression the filter expression
+     * Add another filter provider to the group
+     * \param provider The provider to add
+     * \return itself
      */
-    void setFilter( const QgsVectorLayer *layer, const QgsExpression &expression );
+    QgsFeatureFilterProviderGroup &addProvider( const QgsFeatureFilterProvider *provider );
+
 
   private:
-    QMap<QString, QString> mFilters;
+    QList<const QgsFeatureFilterProvider *> mProviders;
 };
 
 #endif
