+++ src/server/CMakeLists.txt
@@ -46,9 +46,7 @@ set(QGIS_SERVER_SRCS
   qgsservice.cpp
   qgsservicenativeloader.cpp
   qgsserviceregistry.cpp
-  qgsfeaturefilterprovidergroup.cpp
   qgsdimensionfilter.cpp
-  qgsfeaturefilter.cpp
   qgsstorebadlayerinfo.cpp
   qgsserverquerystringparameter.cpp
 )
